+++ src/main/java/com/example/cns/projectPost/domain/ProjectPost.java
@@ -1,8 +1,8 @@
-package com.example.cns.projectPost.projectPost.domain;
+package com.example.cns.projectPost.domain;
 
 import com.example.cns.member.domain.Member;
 import com.example.cns.project.domain.Project;
-import com.example.cns.projectPost.projectPost.type.OpinionType;
+import com.example.cns.projectPost.type.OpinionType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
