+++ src/main/java/com/example/cns/projectPost/domain/ProjectPostOpinion.java
@@ -1,7 +1,7 @@
-package com.example.cns.projectPost.projectPost.domain;
+package com.example.cns.projectPost.domain;
 
 import com.example.cns.member.domain.Member;
-import com.example.cns.projectPost.projectPost.type.OpinionType;
+import com.example.cns.projectPost.type.OpinionType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
