+++ src/main/java/com/example/cns/projectPost/dto/request/ProjectPostRequest.java
@@ -1,8 +1,8 @@
-package com.example.cns.projectPost.projectPost.dto.request;
+package com.example.cns.projectPost.dto.request;
 
 import com.example.cns.member.domain.Member;
 import com.example.cns.project.domain.Project;
-import com.example.cns.projectPost.projectPost.domain.ProjectPost;
+import com.example.cns.projectPost.domain.ProjectPost;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
