--- src/main/java/com/example/cns/projectPost/projectPost/type/OpinionType.java
@@ -1,5 +1,0 @@
-package com.example.cns.projectPost.projectPost.type;
-
-public enum OpinionType {
-    PROS,CONS,CHECK
-}
