+++ src/main/java/com/example/cns/projectPost/service/ProjectPostService.java
@@ -1,4 +1,4 @@
-package com.example.cns.projectPost.projectPost.service;
+package com.example.cns.projectPost.service;
 
 import com.example.cns.common.exception.BusinessException;
 import com.example.cns.common.exception.ExceptionCode;
@@ -6,15 +6,15 @@ import com.example.cns.member.domain.Member;
 import com.example.cns.member.domain.repository.MemberRepository;
 import com.example.cns.project.domain.Project;
 import com.example.cns.project.domain.repository.ProjectRepository;
-import com.example.cns.projectPost.projectPost.domain.ProjectPost;
-import com.example.cns.projectPost.projectPost.domain.ProjectPostOpinion;
-import com.example.cns.projectPost.projectPost.domain.repository.ProjectPostListRepositoryImpl;
-import com.example.cns.projectPost.projectPost.domain.repository.ProjectPostOpinionRepository;
-import com.example.cns.projectPost.projectPost.domain.repository.ProjectPostRepository;
-import com.example.cns.projectPost.projectPost.dto.request.ProjectPostOpinionRequest;
-import com.example.cns.projectPost.projectPost.dto.request.ProjectPostRequest;
-import com.example.cns.projectPost.projectPost.dto.response.ProjectPostResponse;
-import com.example.cns.projectPost.projectPost.type.OpinionType;
+import com.example.cns.projectPost.domain.ProjectPost;
+import com.example.cns.projectPost.domain.ProjectPostOpinion;
+import com.example.cns.projectPost.domain.repository.ProjectPostListRepositoryImpl;
+import com.example.cns.projectPost.domain.repository.ProjectPostOpinionRepository;
+import com.example.cns.projectPost.domain.repository.ProjectPostRepository;
+import com.example.cns.projectPost.dto.request.ProjectPostOpinionRequest;
+import com.example.cns.projectPost.dto.request.ProjectPostRequest;
+import com.example.cns.projectPost.dto.response.ProjectPostResponse;
+import com.example.cns.projectPost.type.OpinionType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
