+++ src/main/java/com/example/cns/projectPost/type/OpinionType.java
@@ -1,0 +1,5 @@
+package com.example.cns.projectPost.type;
+
+public enum OpinionType {
+    PROS,CONS,CHECK
+}
